Include a Backtrace (new class) in CqlExceptions thrown at runtime
A backtrace represents a stack of call-like frames from the root of an
evaluation to a particular sub-expression, commonly a sub-expression
in which an exception was thrown. Client code can extract this
Backtrace object from a caught exception and use it to display a
backtrace similar to those known from other programming languages.