Remove most profiling overhead
Before this change, there were two performance issues in the profile
construction code

1. Profile.register traversed the whole stack to find the appropriate
   profile node which is O(N) where N is the depth of the "call stack".

2. Children of profile nodes were stored in a HashMap the keys of
   which relied on object-equality involving Element
   objects. Object-equality for those is very expensive.

This change solves issue 1 by remembering the stack of profile nodes
that correspond to the stack of current activation frames and
incrementally updating the top of the stack which is O(1).

This change solves issue 2 by (basically; the context adds a slight
complication) storing the children of profile nodes in an
IdentityHashMap so that Element.equals is not used.